requirements.txt
Django==1.3.1
distribute==0.6.19
django-tastypie==0.9.10
mimeparse==0.1.3
psycopg2==2.4.2
python-dateutil==1.5
python-digest==1.7
wsgiref==0.1.2
